pyproject.toml
[tool.poetry]
name = "16.09.23-proxy-checker"
version = "0.1.0"
description = ""
authors = ["Ariulis <andrey59995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests-html = "^0.10.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
